apps/api/requirements.txt
fastapi==0.115.6
uvicorn==0.32.0
python-multipart
opencv-python-headless==0.0.20
pillow==12.1.0
numpy==2.4.2
opencv-python==4.10.0
requests==2.32.5
httpx==0.28.1
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
aiofiles==24.11.0
openmeteo-requests==1.1.0
requests-cache==1.2.3
retry-requests==1.1.0
pandas==3.0.0
azure-cognitiveservices-vision-customvision
msrest
